fix: sua loi nop CV, phan quyen CandidateProfile va hien thi badge trang thai
- Fix JobDetailController: doi userId tu int sang long tranh overflow
- Fix UserCVDAO.getCVsByUserId: doi setInt sang setLong nhat quan kieu du lieu
- Fix CandidateDAO.createProfile: them EmploymentStatus, IsPublicProfile vao INSERT
  de tranh loi constraint NOT NULL khi dang ky tai khoan moi
- Fix UserDAO.findAllCandidates: lay CV moi nhat tu UserCVs thay vi CandidateResumes
- Fix employer_candidate_discovery.jsp: them contextPath vao href Xem CV
- Fix manage_cv.jsp: sua badge trang thai dung inline style thay vi bg-opacity-15
  khong hop le trong Bootstrap 5

+++ HireHubProject/src/java/controller/JobDetailController.java
@@ -99,7 +99,8 @@ public class JobDetailController extends HttpServlet {
                 Object userIdObj = session.getAttribute("userId");
                 if (userIdObj != null) {
                     try {
-                        int userId = Integer.parseInt(String.valueOf(userIdObj));
+                        // Dùng long thay vì int để tránh overflow khi userId lớn
+                        long userId = Long.parseLong(String.valueOf(userIdObj));
                         dal.UserCVDAO cvDAO = new dal.UserCVDAO();
                         java.util.List<model.UserCV> userCVList = cvDAO.getCVsByUserId(userId);
                         request.setAttribute("userCVList", userCVList);
